requirements.txt
# Add your requirements here like:
# numpy
# scipy>=0.9
python-gitlab>=1.3.0
